pyproject.toml
[tool.poetry]
name = "browserpilot"
version = "0.2.53"
description = "Natural language browser automation"
authors = ["Andrew Han <handrew11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
click = "^8.1.7"
python = "^3.10"
openai = "^1.55.0"
selenium = "^4.26.1"
tqdm = "^4.67.0"
beautifulsoup4 = "^4.12.3"
pyyaml = "^6.0.2"
langchain = "^0.1.20"
llama-index = "^0.10.68"
langchain_openai = "^0.1.7"
python-dotenv = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
